tech(api): rename usecase and service
